Update BERT model file structure
