+++ keras_nlp/models/bert/bert_models.py
@@ -12,192 +12,22 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-"""BERT model configurable class and preconfigured versions."""
+"""BERT backbone models."""
 
 import os
-from collections import defaultdict
 
 import tensorflow as tf
 from tensorflow import keras
 
-from keras_nlp.layers.multi_segment_packer import MultiSegmentPacker
 from keras_nlp.layers.position_embedding import PositionEmbedding
 from keras_nlp.layers.transformer_encoder import TransformerEncoder
-from keras_nlp.tokenizers.word_piece_tokenizer import WordPieceTokenizer
+from keras_nlp.models.bert.bert_checkpoints import checkpoints
+from keras_nlp.models.bert.bert_checkpoints import compatible_checkpoints
+from keras_nlp.models.bert.bert_checkpoints import vocabularies
 
 
 def bert_kernel_initializer(stddev=0.02):
     return keras.initializers.TruncatedNormal(stddev=stddev)
-
-
-# TODO(jbischof): document checkpoints in keras.io and use URL in docstrings
-# Metadata for loading pretrained model weights.
-checkpoints = {
-    "bert_tiny_uncased_en": {
-        "model": "BertTiny",
-        "vocabulary": "uncased_en",
-        "description": (
-            "Tiny size of BERT where all input is lowercased. "
-            "Trained on English Wikipedia + BooksCorpus."
-        ),
-        "weights_url": "https://storage.googleapis.com/keras-nlp/models/bert_tiny_uncased_en/model.h5",
-        "weights_hash": "c2b29fcbf8f814a0812e4ab89ef5c068",
-    },
-    "bert_small_uncased_en": {
-        "model": "BertSmall",
-        "vocabulary": "uncased_en",
-        "description": (
-            "Small size of BERT where all input is lowercased. "
-            "Trained on English Wikipedia + BooksCorpus."
-        ),
-        "weights_url": "https://storage.googleapis.com/keras-nlp/models/bert_small_uncased_en/model.h5",
-        "weights_hash": "08632c9479b034f342ba2c2b7afba5f7",
-    },
-    "bert_medium_uncased_en": {
-        "model": "BertMedium",
-        "vocabulary": "uncased_en",
-        "description": (
-            "Medium size of BERT where all input is lowercased. "
-            "Trained on English Wikipedia + BooksCorpus."
-        ),
-        "weights_url": "https://storage.googleapis.com/keras-nlp/models/bert_medium_uncased_en/model.h5",
-        "weights_hash": "bb990e1184ec6b6185450c73833cd661",
-    },
-    "bert_base_uncased_en": {
-        "model": "BertBase",
-        "vocabulary": "uncased_en",
-        "description": (
-            "Base size of BERT where all input is lowercased. "
-            "Trained on English Wikipedia + BooksCorpus."
-        ),
-        "weights_url": "https://storage.googleapis.com/keras-nlp/models/bert_base_uncased_en/model.h5",
-        "weights_hash": "9b2b2139f221988759ac9cdd17050b31",
-    },
-    "bert_base_cased_en": {
-        "model": "BertBase",
-        "vocabulary": "cased_en",
-        "description": (
-            "Base size of BERT where case is maintained. "
-            "Trained on English Wikipedia + BooksCorpus."
-        ),
-        "weights_url": "https://storage.googleapis.com/keras-nlp/models/bert_base_cased_en/model.h5",
-        "weights_hash": "f94a6cb012e18f4fb8ec92abb91864e9",
-    },
-    "bert_base_zh": {
-        "model": "BertBase",
-        "vocabulary": "zh",
-        "description": ("Base size of BERT. Trained on Chinese Wikipedia."),
-        "weights_url": "https://storage.googleapis.com/keras-nlp/models/bert_base_zh/model.h5",
-        "weights_hash": "79afa421e386076e62ab42dad555ab0c",
-    },
-    "bert_base_multi_cased": {
-        "model": "BertBase",
-        "vocabulary": "multi_cased",
-        "description": ("Base size of BERT. Trained on Chinese Wikipedia."),
-        "weights_url": "https://storage.googleapis.com/keras-nlp/models/bert_base_multi_cased/model.h5",
-        "weights_hash": "b0631cec0a1f2513c6cfd75ba29c33aa",
-    },
-    "bert_large_uncased_en": {
-        "model": "BertLarge",
-        "vocabulary": "uncased_en",
-        "description": (
-            "Large size of BERT where all input is lowercased. "
-            "Trained on English Wikipedia + BooksCorpus."
-        ),
-        "weights_url": "https://storage.googleapis.com/keras-nlp/models/bert_large_uncased_en/model.h5",
-        "weights_hash": "cc5cacc9565ef400ee4376105f40ddae",
-    },
-    "bert_large_cased_en": {
-        "model": "BertLarge",
-        "vocabulary": "cased_en",
-        "description": (
-            "Base size of BERT where case is maintained. "
-            "Trained on English Wikipedia + BooksCorpus."
-        ),
-        "weights_url": "https://storage.googleapis.com/keras-nlp/models/bert_large_cased_en/model.h5",
-        "weights_hash": "8b8ab82290bbf4f8db87d4f100648890",
-    },
-}
-
-
-# Index checkpoints by arch compatibility and create lookup function
-checkpoints_per_arch = defaultdict(set)
-for arch, metadata in checkpoints.items():
-    checkpoints_per_arch[metadata["model"]].add(arch)
-
-
-def compatible_checkpoints(arch):
-    """Returns a list of compatible checkpoints per arch"""
-    return checkpoints_per_arch[arch]
-
-
-# Metadata for loading pretrained tokenizer vocabularies.
-# We need the vocabulary_size hardcoded so we can instantiate a BERT network
-# with the right embedding size without downloading the matching vocabulary.
-# TODO(mattdangerw): Update our bucket structure so the vocabularies are
-# stored in an independent way, rather than reading from the base model.
-vocabularies = {
-    "uncased_en": {
-        "description": (
-            "The vocabulary for BERT models trained on "
-            "English Wikipedia + BooksCorpus where case is discarded."
-        ),
-        "vocabulary_url": "https://storage.googleapis.com/keras-nlp/models/bert_base_uncased_en/vocab.txt",
-        "vocabulary_hash": "64800d5d8528ce344256daf115d4965e",
-        "vocabulary_size": 30522,
-        "lowercase": True,
-    },
-    "cased_en": {
-        "description": (
-            "The vocabulary for BERT models trained on "
-            "English Wikipedia + BooksCorpus where case is maintained."
-        ),
-        "vocabulary_url": "https://storage.googleapis.com/keras-nlp/models/bert_base_cased_en/vocab.txt",
-        "vocabulary_hash": "bb6ca9b42e790e5cd986bbb16444d0e0",
-        "vocabulary_size": 28996,
-        "lowercase": False,
-    },
-    "zh": {
-        "description": (
-            "The vocabulary for BERT models trained on Chinese Wikipedia."
-        ),
-        "vocabulary_url": "https://storage.googleapis.com/keras-nlp/models/bert_base_zh/vocab.txt",
-        "vocabulary_hash": "3b5b76c4aef48ecf8cb3abaafe960f09",
-        "vocabulary_size": 21128,
-        "lowercase": False,
-    },
-    "multi_cased": {
-        "description": (
-            "The vocabulary for BERT models trained on trained on Wikipedias "
-            "of 104 languages."
-        ),
-        "vocabulary_url": "https://storage.googleapis.com/keras-nlp/models/bert_base_multi_cased/vocab.txt",
-        "vocabulary_hash": "d9d865138d17f1958502ed060ecfeeb6",
-        "vocabulary_size": 119547,
-        "lowercase": False,
-    },
-}
-
-
-def _handle_pretrained_tokenizer_arguments(vocabulary, lowercase):
-    """Look up pretrained defaults for tokenizer arguments.
-
-    This helper will validate the `vocabulary` and `lowercase` arguments, and
-    fully resolve them in the case we are loading pretrained weights.
-    """
-
-    # TODO(jbischof): allow user to lookup vocabulary from checkpoint name
-    if isinstance(vocabulary, str) and vocabulary in vocabularies:
-        metadata = vocabularies[vocabulary]
-        vocabulary = keras.utils.get_file(
-            "vocab.txt",
-            metadata["vocabulary_url"],
-            cache_subdir=os.path.join("models", "bert", vocabulary),
-            file_hash=metadata["vocabulary_hash"],
-        )
-        lowercase = metadata["lowercase"]
-
-    return vocabulary, lowercase
 
 
 def _handle_pretrained_model_arguments(bert_variant, weights, vocabulary_size):
@@ -433,155 +263,6 @@ class BertCustom(keras.Model):
             }
         )
         return config
-
-
-PREPROCESSOR_DOCSTRING = """BERT preprocessor with pretrained vocabularies.
-
-This preprocessing layer will do three things:
-
- - Tokenize any number of inputs using a
-   `keras_nlp.tokenizers.WordPieceTokenizer`.
- - Pack the inputs together using a `keras_nlp.layers.MultiSegmentPacker`.
-   with the appropriate `"[CLS]"`, `"[SEP]"` and `"[PAD]"` tokens.
- - Construct a dictionary of with keys `"token_ids"`, `"segment_ids"`,
-   `"padding_mask"`, that can be passed directly to a BERT model.
-
-This layer will accept either a tuple of (possibly batched) inputs, or a single
-input tensor. If a single tensor is passed, it will be packed equivalently to
-a tuple with a single element.
-
-The WordPiece tokenizer can be accessed via the `tokenizer` property on this
-layer, and can be used directly for custom packing on inputs.
-
-Args:
-    vocabulary: One of a list of vocabulary terms, a vocabulary filename, or
-        the name of the pretrained vocabulary. For a pretrained vocabulary,
-        `vocabulary` should be one of {names}, and should match the `weights`
-        parameter of any pretrained BERT model.
-    lowercase: If `True`, input will be lowercase before tokenization. If
-        `vocabulary` is set to a pretrained vocabulary, this parameter will
-        be inferred.
-    sequence_length: The length of the packed inputs. Only used if
-        `pack_inputs` is True.
-    truncate: The algorithm to truncate a list of batched segments to fit
-        within `sequence_length`. Only used if
-        `pack_inputs` is True. The value can be either `round_robin` or
-        `waterfall`:
-            - `"round_robin"`: Available space is assigned one token at a
-                time in a round-robin fashion to the inputs that still need
-                some, until the limit is reached.
-            - `"waterfall"`: The allocation of the budget is done using a
-                "waterfall" algorithm that allocates quota in a
-                left-to-right manner and fills up the buckets until we run
-                out of budget. It supports an arbitrary number of segments.
-
-Examples:
-```python
-preprocessor = keras_nlp.models.BertPreprocessor(vocabulary="uncased_en")
-
-# Tokenize and pack a single sentence directly.
-preprocessor("The quick brown fox jumped.")
-
-# Tokenize and pack a multiple sentence directly.
-preprocessor(("The quick brown fox jumped.", "Call me Ishmael."))
-
-# Map a dataset to preprocess a single sentence.
-features = ["The quick brown fox jumped.", "I forgot my homework."]
-labels = [0, 1]
-ds = tf.data.Dataset.from_tensor_slices((features, labels))
-ds = ds.map(
-    lambda x, y: (preprocessor(x), y),
-    num_parallel_calls=tf.data.AUTOTUNE,
-)
-
-# Map a dataset to preprocess a multiple sentences.
-first_sentences = ["The quick brown fox jumped.", "Call me Ishmael."]
-second_sentences = ["The fox tripped.", "Oh look, a whale."]
-labels = [1, 1]
-ds = tf.data.Dataset.from_tensor_slices(((first_sentences, second_sentences), labels))
-ds = ds.map(
-    lambda x, y: (preprocessor(x), y),
-    num_parallel_calls=tf.data.AUTOTUNE,
-)
-```
-"""
-
-
-class BertPreprocessor(keras.layers.Layer):
-    def __init__(
-        self,
-        vocabulary="uncased_en",
-        lowercase=False,
-        sequence_length=512,
-        truncate="round_robin",
-        **kwargs,
-    ):
-        super().__init__(**kwargs)
-
-        vocabulary, lowercase = _handle_pretrained_tokenizer_arguments(
-            vocabulary, lowercase
-        )
-
-        self.tokenizer = WordPieceTokenizer(
-            vocabulary=vocabulary,
-            lowercase=lowercase,
-        )
-
-        # Check for necessary special tokens.
-        cls_token = "[CLS]"
-        sep_token = "[SEP]"
-        pad_token = "[PAD]"
-        for token in [cls_token, pad_token, sep_token]:
-            if token not in self.tokenizer.get_vocabulary():
-                raise ValueError(
-                    f"Cannot find token `'{token}'` in the provided "
-                    f"`vocabulary`. Please provide `'{token}'` in your "
-                    "`vocabulary` or use a pretrained `vocabulary` name."
-                )
-
-        self.pad_token_id = self.tokenizer.token_to_id(pad_token)
-        self.packer = MultiSegmentPacker(
-            start_value=self.tokenizer.token_to_id(cls_token),
-            end_value=self.tokenizer.token_to_id(sep_token),
-            pad_value=self.pad_token_id,
-            truncate=truncate,
-            sequence_length=sequence_length,
-        )
-
-    def vocabulary_size(self):
-        """Returns the vocabulary size of the tokenizer."""
-        return self.tokenizer.vocabulary_size()
-
-    def get_config(self):
-        config = super().get_config()
-        config.update(
-            {
-                "vocabulary": self.tokenizer.vocabulary,
-                "lowercase": self.tokenizer.lowercase,
-                "sequence_length": self.packer.sequence_length,
-                "truncate": self.packer.truncate,
-            }
-        )
-        return config
-
-    def call(self, inputs):
-        if not isinstance(inputs, (list, tuple)):
-            inputs = [inputs]
-
-        inputs = [self.tokenizer(x) for x in inputs]
-        token_ids, segment_ids = self.packer(inputs)
-        return {
-            "token_ids": token_ids,
-            "segment_ids": segment_ids,
-            "padding_mask": token_ids != self.pad_token_id,
-        }
-
-
-setattr(
-    BertPreprocessor,
-    "__doc__",
-    PREPROCESSOR_DOCSTRING.format(names=", ".join(vocabularies)),
-)
 
 
 MODEL_DOCSTRING = """Bert "{type}" architecture.
