+++ keras_nlp/network_tests/test_load_ckpts.py
@@ -16,20 +16,18 @@
 import pytest
 import tensorflow as tf
 
-import keras_nlp
-from keras_nlp.models.bert import checkpoints as bert_checkpoints
-from keras_nlp.models.bert import (
-    model_class_by_name as bert_model_class_by_name,
-)
-from keras_nlp.models.bert import vocabularies as bert_vocabularies
+from keras_nlp.models.bert import bert_checkpoints
+from keras_nlp.models.bert import bert_models
+from keras_nlp.models.bert import bert_preprocessing
+from keras_nlp.models.bert import bert_tasks
 
 
 @pytest.mark.slow
 class BertCkptTest(tf.test.TestCase):
     def test_load_bert(self):
-        for checkpoint in bert_checkpoints:
-            bert_variant = bert_checkpoints[checkpoint]["model"]
-            bert_class = bert_model_class_by_name(bert_variant)
+        for checkpoint in bert_checkpoints.checkpoints:
+            bert_variant = bert_checkpoints.checkpoints[checkpoint]["model"]
+            bert_class = bert_models.model_class_by_name(bert_variant)
             model = bert_class(weights=checkpoint)
             input_data = {
                 "token_ids": tf.random.uniform(
@@ -43,8 +41,8 @@ class BertCkptTest(tf.test.TestCase):
             model(input_data)
 
     def test_load_bert_backbone_string(self):
-        for checkpoint in bert_checkpoints:
-            classifier = keras_nlp.models.bert_tasks.BertClassifier(
+        for checkpoint in bert_checkpoints.checkpoints:
+            classifier = bert_tasks.BertClassifier(
                 checkpoint, 4, name="classifier"
             )
             input_data = {
@@ -61,7 +59,7 @@ class BertCkptTest(tf.test.TestCase):
             classifier(input_data)
 
     def test_classifier_default_args(self):
-        classifier = keras_nlp.models.bert_tasks.BertClassifier()
+        classifier = bert_tasks.BertClassifier()
         input_data = {
             "token_ids": tf.random.uniform(
                 shape=(1, 512),
@@ -74,6 +72,8 @@ class BertCkptTest(tf.test.TestCase):
         classifier(input_data)
 
     def test_load_vocabularies(self):
-        for vocabulary in bert_vocabularies:
-            tokenizer = keras_nlp.models.BertPreprocessor(vocabulary=vocabulary)
+        for vocabulary in bert_checkpoints.vocabularies:
+            tokenizer = bert_preprocessing.BertPreprocessor(
+                vocabulary=vocabulary,
+            )
             tokenizer("The quick brown fox.")
